MAR-2450. Created mixin for reuse the copy function in components, updated snapshots
